Conflicts:
	conf/application.conf

+++ app/models/Competence.java
@@ -1,8 +1,6 @@
 package models;
 
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
@@ -13,30 +11,39 @@ import play.db.ebean.Model;
 import java.util.List;
 import java.util.ArrayList;
 
+/**
+ * Competence entity managed by Ebean
+ */
 @Entity
 @Table(name = "competence")
 public class Competence extends Model {
 	
 	@Id
-	public int competence_id;
+	public long competence_id;
 
 	@Constraints.Required(message="Required")
 	@Formats.NonEmpty
 	public String name;
 
+    /**
+     * Generic query helper for entity Competence by name
+     */
 	public static Model.Finder<String, Competence> find = new Model.Finder(String.class, Competence.class);
 
     /**
      * Return a list of all Competences.
      * 
-     * @return List of all Competences.
+     * @return List <Competence>
      */
 	public static List<Competence> getAll() {
 		return find.all();
 	}
 
-	public static Competence findById(int id) {
-                return find.where().eq("competence_id", id).findUnique();
+    /**
+     * Generic query helper for entity Competence with id Long
+     */
+	public static Competence findById(long id) {
+		return find.where().eq("competence_id", id).findUnique();
 	}
 
 	@Override
@@ -44,10 +51,15 @@ public class Competence extends Model {
 		return name;
 	}
 
+    /**
+     * Return a list of Strings with the names of all Competences.
+     * 
+     * @return List <Competence>
+     */
 	public static List<String> getList() {
 		List<String> ret = new ArrayList<String>();
 		for(Competence comp: getAll()) {
-			ret.add(Integer.toString(comp.competence_id));
+			ret.add(Long.toString(comp.competence_id));
 		}
 		return ret;
 	}
